Implementing routes

+++ models/movie.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, Float, Integer, String
-from model import Base
+from models import Base
 
 class Movie(Base):
     __tablename__ = 'movie'
@@ -8,7 +8,7 @@ class Movie(Base):
     title = Column(String(100), unique=True)
     poster_url = Column(String(2100))
     running_time = Column(Integer, nullable=False)
-    budget = Column(Integer, nullable=False)
+    budget = Column(Float, nullable=False)
     box_office = Column(Float)
 
     def __init__(self, title:str, poster_url:str,
